Гуков Арман задание 9/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0 